+++ pkg_thunderstore/goatgirl-DynamicZoom/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 1.0.1
+- Fixed an issue where the dynamic zoom ignored any obstructions
+    - Can be re-ignored in the config, if you so please ("Ignore Obstructions")
 ## 1.0.0
 - Initial release